refactor: move workerpool to `internal` package
