+++ internal/workerpool_test.go
@@ -1,4 +1,4 @@
-package scaleway
+package internal
 
 import (
 	"errors"
